+++ src/cwl_platform/arvados_platform.py
@@ -374,6 +374,11 @@ class ArvadosPlatform(Platform):
 
         return None
 
+    def get_task_outputs(self, task):
+        ''' Return a list of output fields of the task '''
+        cwl_output = self._load_cwl_output(task)
+        return list(cwl_output.keys())
+
     def get_task_output_filename(self, task: ArvadosTask, output_name):
         ''' Retrieve the output field of the task and return filename'''
         cwl_output_collection = arvados.collection.Collection(task.container_request['output_uuid'],
@@ -435,6 +440,34 @@ class ArvadosPlatform(Platform):
         if len(search_result['items']) > 0:
             return search_result['items'][0]
         return None
+
+    def rename_file(self, fileid, new_filename):
+        '''
+        Rename a file to new_filename.
+
+        :param file: File ID to rename
+        :param new_filename: str of new filename
+        '''
+        collection_uuid = fileid.split('keep:')[1].split('/')[0]
+        filepath = fileid.split(collection_uuid+'/')[1]
+        if len(filepath.split('/'))>1:
+            newpath = '/'.join(filepath.split('/')[:-1]+[new_filename])
+        else:
+            newpath = new_filename
+        collection = arvados.collection.Collection(collection_uuid, api_client=self.api)
+        collection.copy(filepath, newpath)
+        collection.remove(filepath, recursive=True)
+        collection.save()
+
+    def roll_file(self, project, file_name):
+        '''
+        Roll (find and rename) a file in a project.
+
+        :param project: The project the file is located in
+        :param file_name: The filename that needs to be rolled
+        '''
+        # Each run of a workflow will have a unique output collection, hence there will be no
+        # name conflicts.
 
     def stage_output_files(self, project, output_files):
         '''