+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.2](https://github.com/NGS360/PAML/releases/tag/v0.2) - 2024-06-14
+
+### Added
+
+- get_task_outputs method
+- rename_output_files
+
+### Fixed
+
+- Fixed bug in how project is retrieved from SevenBridges
+
 ## [v0.1.1](https://github.com/NGS360/PAML/releases/tag/v0.1.1) - 2024-04-09
 
 ### Fixed